# cirras-claims-ngclient-lib/cirras-claims-ngclient-lib/src/main/angular/src/swagger/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "CIRRAS Claims V1 API",
    "version" : "1.0.0-SNAPSHOT"
  },
  "paths" : {
    "/claims/{insuranceClaimGuid}" : {
      "get" : {
        "summary" : "Get the claim",
        "operationId" : "Get the claim.",
        "parameters" : [ {
          "name" : "RequestId",
          "in" : "header",
          "description" : "The unique identity of the request assigned by the originator to help trace the request through the logs.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Rest-Version",
          "in" : "header",
          "description" : "The version of the Rest API supported by the requesting client.",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "Cache-Control",
          "in" : "header",
          "description" : "Used to specify directives for caches along the request-response chain",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Pragma",
          "in" : "header",
          "description" : "An implementation-specific header that may have various effects along the request-response chain. It is used for backwards compatibility with HTTP/1.0 caches where the Cache-Control HTTP/1.1 header is not yet present.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Contains the credentials to authenticate a user agent with a server.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "insuranceClaimGuid",
          "in" : "path",
          "description" : "The GUID of the claim.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "headers" : {
              "ETag" : {
                "description" : "The ETag response-header field provides the current value of the entity tag for the requested variant.",
                "style" : "simple"
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InsuranceClaimRsrc"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InsuranceClaimRsrc"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListRsrc"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListRsrc"
                }
              }
            }
          }
        },
        "security" : [ {
          "Webade-OAUTH2" : [ "CIRRAS_CLAIMS.GET_CLAIM" ]
        } ],
        "x-auth-type" : "#{wso2.x-auth-type.app_and_app_user}",
        "x-throttling-tier" : "Unlimited"
      },
      "put" : {
        "summary" : "Update claim",
        "operationId" : "Update claim",
        "parameters" : [ {
          "name" : "RequestId",
          "in" : "header",
          "description" : "The unique identity of the request assigned by the originator to help trace the request through the logs.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Rest-Version",
          "in" : "header",
          "description" : "The version of the Rest API supported by the requesting client.",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "Cache-Control",
          "in" : "header",
          "description" : "Used to specify directives for caches along the request-response chain",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Pragma",
          "in" : "header",
          "description" : "An implementation-specific header that may have various effects along the request-response chain. It is used for backwards compatibility with HTTP/1.0 caches where the Cache-Control HTTP/1.1 header is not yet present.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Contains the credentials to authenticate a user agent with a server.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "description" : "The If-Match request-header must match the current eTag of the resource or the request will fail.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "insuranceClaimGuid",
          "in" : "path",
          "description" : "The GUID of the claim resource.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "The claim  resource containing the new values.",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/InsuranceClaimRsrc"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/InsuranceClaimRsrc"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "headers" : {
              "ETag" : {
                "description" : "The ETag response-header field provides the current value of the entity tag for the requested variant.",
                "style" : "simple"
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InsuranceClaimRsrc"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InsuranceClaimRsrc"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListRsrc"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListRsrc"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "409" : {
            "description" : "Conflict"
          },
          "412" : {
            "description" : "Precondition Failed"
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListRsrc"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListRsrc"
                }
              }
            }
          }
        },
        "security" : [ {
          "Webade-OAUTH2" : [ "CIRRAS_CLAIMS.UPDATE_CLAIM" ]
        } ],
        "x-auth-type" : "#{wso2.x-auth-type.app_and_app_user}",
        "x-throttling-tier" : "Unlimited"
      },
      "delete" : {
        "summary" : "Delete claim",
        "operationId" : "Delete claim",
        "parameters" : [ {
          "name" : "RequestId",
          "in" : "header",
          "description" : "The unique identity of the request assigned by the originator to help trace the request through the logs.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Rest-Version",
          "in" : "header",
          "description" : "The version of the Rest API supported by the requesting client.",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "Cache-Control",
          "in" : "header",
          "description" : "Used to specify directives for caches along the request-response chain",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Pragma",
          "in" : "header",
          "description" : "An implementation-specific header that may have various effects along the request-response chain. It is used for backwards compatibility with HTTP/1.0 caches where the Cache-Control HTTP/1.1 header is not yet present.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Contains the credentials to authenticate a user agent with a server.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "description" : "The If-Match request-header must match the current eTag of the resource or the request will fail.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "insuranceClaimGuid",
          "in" : "path",
          "description" : "The GUID of the claim resource.",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "409" : {
            "description" : "Conflict"
          },
          "412" : {
            "description" : "Precondition Failed"
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListRsrc"
                }
              }
            }
          }
        },
        "security" : [ {
          "Webade-OAUTH2" : [ "CIRRAS_CLAIMS.DELETE_CLAIM" ]
        } ],
        "x-auth-type" : "#{wso2.x-auth-type.app_and_app_user}",
        "x-throttling-tier" : "Unlimited"
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "InsuranceClaimRsrc" : {
        "required" : [ "@type" ],
        "type" : "object",
        "properties" : {
          "links" : {
            "type" : "array",
            "xml" : {
              "wrapped" : true
            },
            "items" : {
              "$ref" : "#/components/schemas/RelLink"
            }
          },
          "insuranceClaimGuid" : {
            "type" : "string"
          },
          "growerNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "growerName" : {
            "type" : "string"
          },
          "growerAddressLine1" : {
            "type" : "string"
          },
          "growerAddressLine2" : {
            "type" : "string"
          },
          "growerPostalCode" : {
            "type" : "string"
          },
          "insurancePlanCode" : {
            "type" : "string"
          },
          "cropYear" : {
            "type" : "integer",
            "format" : "int32"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "insuranceCoverageCode" : {
            "type" : "string"
          },
          "cropCommodityCode" : {
            "type" : "string"
          },
          "insurableValueSelected" : {
            "type" : "number",
            "format" : "double"
          },
          "insurableValueComplete" : {
            "type" : "number",
            "format" : "double"
          },
          "coverageAmount" : {
            "type" : "number",
            "format" : "double"
          },
          "coverageAmountAssessed" : {
            "type" : "number",
            "format" : "double"
          },
          "coverageAssessedReason" : {
            "type" : "string"
          },
          "coverageAmountAdjusted" : {
            "type" : "number",
            "format" : "double"
          },
          "claimNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "claimVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "primaryPerilCode" : {
            "type" : "string"
          },
          "secondaryPerilCode" : {
            "type" : "string"
          },
          "claimStatusCode" : {
            "type" : "string"
          },
          "totalIncome" : {
            "type" : "number",
            "format" : "double"
          },
          "totalClaimAmount" : {
            "type" : "number",
            "format" : "double"
          },
          "recommendedByUserid" : {
            "type" : "string"
          },
          "recommendedByName" : {
            "type" : "string"
          },
          "recommendedByDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "varieties" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/InsuranceClaimVariety"
            }
          },
          "etag" : {
            "type" : "string",
            "writeOnly" : true
          },
          "@type" : {
            "type" : "string"
          }
        },
        "xml" : {
          "name" : "insuranceClaim",
          "namespace" : "http://claims.cirras.mal.gov.bc.ca/v1/"
        },
        "discriminator" : {
          "propertyName" : "@type"
        }
      },
      "InsuranceClaimVariety" : {
        "type" : "object",
        "properties" : {
          "insuranceClaimGuid" : {
            "type" : "string"
          },
          "insuranceClaimVarietyGuid" : {
            "type" : "string"
          },
          "cropVarietyCode" : {
            "type" : "string"
          },
          "firstPlantedYear" : {
            "type" : "integer",
            "format" : "int32"
          },
          "insurableValueComplete" : {
            "type" : "number",
            "format" : "double"
          },
          "yieldProjected" : {
            "type" : "number",
            "format" : "double"
          },
          "contractedPrice1" : {
            "type" : "number",
            "format" : "double"
          },
          "contractedPrice2" : {
            "type" : "number",
            "format" : "double"
          },
          "contractedPrice3" : {
            "type" : "number",
            "format" : "double"
          },
          "contractedPrice4" : {
            "type" : "number",
            "format" : "double"
          },
          "contractedPrice5" : {
            "type" : "number",
            "format" : "double"
          },
          "averagePrice" : {
            "type" : "number",
            "format" : "double"
          },
          "limitedValue" : {
            "type" : "number",
            "format" : "double"
          },
          "insurableValue" : {
            "type" : "number",
            "format" : "double"
          },
          "yieldAssessed" : {
            "type" : "number",
            "format" : "double"
          },
          "yieldTotal" : {
            "type" : "number",
            "format" : "double"
          },
          "yieldActual" : {
            "type" : "number",
            "format" : "double"
          },
          "varietyIncome" : {
            "type" : "number",
            "format" : "double"
          },
          "yieldAssessedReason" : {
            "type" : "string"
          }
        }
      },
      "RelLink" : {
        "required" : [ "@type" ],
        "type" : "object",
        "properties" : {
          "rel" : {
            "type" : "string",
            "xml" : {
              "attribute" : true
            }
          },
          "href" : {
            "type" : "string",
            "xml" : {
              "attribute" : true
            }
          },
          "method" : {
            "type" : "string",
            "xml" : {
              "attribute" : true
            }
          },
          "etag" : {
            "type" : "string",
            "writeOnly" : true
          },
          "@type" : {
            "type" : "string"
          }
        },
        "xml" : {
          "name" : "rellink",
          "namespace" : "http://common.wfone.nrs.gov.bc.ca/v1/"
        },
        "discriminator" : {
          "propertyName" : "@type"
        }
      },
      "MessageListRsrc" : {
        "required" : [ "@type" ],
        "type" : "object",
        "properties" : {
          "messages" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MessageRsrc"
            }
          },
          "etag" : {
            "type" : "string",
            "writeOnly" : true
          },
          "@type" : {
            "type" : "string"
          }
        },
        "xml" : {
          "name" : "messages",
          "namespace" : "http://common.wfone.nrs.gov.bc.ca/v1/"
        },
        "discriminator" : {
          "propertyName" : "@type"
        }
      },
      "MessageRsrc" : {
        "required" : [ "@type" ],
        "type" : "object",
        "properties" : {
          "path" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "messageTemplate" : {
            "type" : "string"
          },
          "messageArguments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "@type" : {
            "type" : "string"
          }
        },
        "xml" : {
          "name" : "message",
          "namespace" : "http://common.wfone.nrs.gov.bc.ca/v1/"
        },
        "discriminator" : {
          "propertyName" : "@type"
        }
      },
...
    }
  }
}